Refactor test case for member-only articles
The change here is in how we obtain the article_id for testing. Previously, we were getting the id of the first article in the database. Now, I've updated this to specifically fetch the id of the first member-only article. This ensures that the test case is using an id that corresponds to a member-only article, which is necessary for the test to pass.